Feature/hotfix

+++ src/main/java/side/mimi/mdd/restApi/Disk/service/DiskService.java
@@ -400,13 +400,13 @@ public class DiskService {
 		}
 		if(files != null && files.length > 4) throw new AppException(ErrorCode.IMG_COUNT_LIMIT, ErrorCode.IMG_COUNT_LIMIT.getMessage());
 
-		if(dto.getIsBookmark() != null && dto.getIsBookmark()){
+		if(dto != null && dto.getIsBookmark() != null && dto.getIsBookmark()){
 			if (bookmarkedDiskList.size() >= 3 && bookmarkedDiskList.stream().noneMatch(bookmarkedDisk -> bookmarkedDisk.getDiskId().equals(myDisk.getDiskId())))
 				throw new AppException(ErrorCode.BOOKMARK_DISK_LIMIT, ErrorCode.BOOKMARK_DISK_LIMIT.getMessage());
 		}
 
 		//이미지 처리
-		if(dto.getDeleteImgList() != null && dto.getDeleteImgList().length > 0) {
+		if(dto != null && dto.getDeleteImgList() != null && dto.getDeleteImgList().length > 0) {
 			for(Long imgId : dto.getDeleteImgList()) {
 				imgRepository.deleteById(imgId);
